login and logout works, adding messaging
