+++ miniproj/src/index.js
@@ -1,21 +1,14 @@
 import React from 'react';
 import './index.css';
 import App from './App';
-import Login from './Login';
 import reportWebVitals from './reportWebVitals';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter, Route, Routes } from "react-router-dom";
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element = {<App />} />
-        <Route path="login" element = {<Login />} />
-      </Routes>
-    </BrowserRouter>
+    <App/>
   </React.StrictMode>
 );
